Backend/api_microservices/fournisseur/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
snowflake-connector-python==3.5.0